Implement (send_)write_command and make write_request(val) use val's type.
At least some devices (at least Xiaomi Mi Band Pulse) differentiate
between write_request and write_command, and error on mismatch.
Functionality already was implemented, this commit exposes it through
BLEGATTStateMachine::send_write_command and Characteristic::write_command.

Characteristic::write_request(uint8_t) was quite arbitrary and unobvious.
At least I expected single-argument version to send uint16_t. This commit
replaces it with templated variant that sends value of any type (not just
numbers) as is. Host is assumed to be Little Endian.
Existing calls should be replaced with write_request(uint8_t(val)).

+++ examples/write.cc
@@ -57,7 +57,7 @@ int main(int argc, char **argv)
 					if(characteristic.uuid == UUID("b8637601-a003-436d-a995-2a7f20bcb3d4"))
 					{
 						//Send a 1 (you can also send longer chunks of data too)
-						characteristic.write_request(1);	
+						characteristic.write_request(uint8_t(1));
 					}
 	};
